+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v1.7.4 (TBA)
 - Added option to set custom greetings and date ([#103](https://github.com/pawelmalak/flame/issues/103))
+- Fallback to web search if local search has zero results ([#129](https://github.com/pawelmalak/flame/issues/129))
 - Added iOS "Add to homescreen" icon ([#131](https://github.com/pawelmalak/flame/issues/131))
 - Added 3 new themes